backend/requirements.txt
Django
django-ninja
django-ninja-jwt
psycopg2-binary
django-cors-headers
# django-auth-ldap
gunicorn
